chore: consolidate more constants
