+++ src/features/defi/providers/idle/components/IdleManager/Deposit/components/Approve.tsx
@@ -20,7 +20,7 @@ import { useWallet } from 'hooks/useWallet/useWallet'
 import type { Asset } from 'lib/asset-service'
 import type { BigNumber } from 'lib/bignumber/bignumber'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
-import { ssRouterContractAddress } from 'lib/investor/investor-idle'
+import { ssRouterContractAddress } from 'lib/investor/constants'
 import { logger } from 'lib/logger'
 import { trackOpportunityEvent } from 'lib/mixpanel/helpers'
 import { MixPanelEvents } from 'lib/mixpanel/types'
