--- src/lib/investor/investor-foxy/constants/enums.ts
@@ -1,3 +1,0 @@
-export enum DefiType {
-  Staking = 'staking',
-}
