+++ src/lib/investor/investor-idle/IdleInvestor.ts
@@ -7,8 +7,9 @@ import filter from 'lodash/filter'
 import Web3 from 'web3'
 import type { Contract } from 'web3-eth-contract'
 import type { Investor } from 'lib/investor'
+import { ssRouterContractAddress } from 'lib/investor/constants'
 import type { IdleVault } from 'lib/investor/investor-idle/constants'
-import { ssRouterAbi, ssRouterContractAddress } from 'lib/investor/investor-idle/constants'
+import { ssRouterAbi } from 'lib/investor/investor-idle/constants'
 import type { PreparedTransaction } from 'lib/investor/investor-idle/IdleOpportunity'
 import { IdleOpportunity } from 'lib/investor/investor-idle/IdleOpportunity'
 import { IdleSdk } from 'lib/investor/investor-idle/IdleSdk'
