+++ src/lib/investor/investor-idle/constants/index.ts
@@ -1,7 +1,6 @@
-export * from 'lib/investor/investor-idle/constants/erc20-abi'
+export * from 'lib/investor/constants/erc20-abi'
 export * from 'lib/investor/investor-idle/constants/idle-cdo-abi'
 export * from 'lib/investor/investor-idle/constants/ss-router-abi'
-export * from 'lib/investor/investor-idle/constants/router-contract'
 export * from 'lib/investor/investor-idle/constants/availableTokens'
 export * from 'lib/investor/investor-idle/constants/idle-token-v4-abi'
 export * from 'lib/investor/investor-idle/constants/idle-strategy-abi'
