+++ src/lib/investor/investor-yearn/constants/index.ts
@@ -1,4 +1,3 @@
-export * from './erc20-abi'
-export * from './router-contract'
+export * from 'lib/investor/constants/router-contract'
 export * from './ss-router-abi'
 export * from './yv2Vaults-abi'
